Define operations in the init function
This gives cleaner code and allows operations to
be defined and registered by specific providers,
such as the dearxiv functionality.
